pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "syndicateclaw"
version = "2.0.0"
description = "Production-oriented agent orchestration platform with stateful graph-based workflows"
requires-python = ">=3.14.3"
license = "MIT"
dependencies = [
    "fastapi>=0.115.14",
    "starlette>=0.49.1",
    "uvicorn[standard]==0.34.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.8.1",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg==0.30.0",
    "alembic==1.15.1",
    "redis[hiredis]==5.2.1",
    "opentelemetry-api==1.30.0",
    "opentelemetry-sdk==1.30.0",
    "opentelemetry-instrumentation-fastapi==0.51b0",
    "opentelemetry-instrumentation-sqlalchemy==0.51b0",
    "opentelemetry-instrumentation-httpx==0.51b0",
    "opentelemetry-exporter-otlp==1.30.0",
    "httpx==0.28.1",
    "structlog==25.1.0",
    "tenacity==9.0.0",
    "python-ulid==3.0.0",
    "pyyaml==6.0.2",
    "jsonschema==4.23.0",
    "packaging==24.2",
    "cryptography==46.0.6",
    "PyJWT[crypto]==2.12.0",
    "passlib[bcrypt]==1.7.4",
    "prometheus-client==0.21.1",
    "jinja2==3.1.6",
    "pytimeparse>=1.1.0",
    "croniter>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "pytest-asyncio==0.25.3",
    "pytest-cov==6.1.1",
    "httpx==0.28.1",
    "asgi-lifespan==2.1.0",
    "factory-boy==3.3.1",
    "ruff==0.11.2",
    "mypy==1.15.0",
    "types-jsonschema==4.23.0.20240813",
    "types-PyYAML==6.0.12.20241230",
    "locust>=2.20.0",
    "pytest-benchmark>=4.0.0",
    "pytest-xdist>=3.5.0",
    "mutmut>=3.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/syndicateclaw"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (require DB/services)",
    "requires_api_keys: marks tests that require real provider API keys",
    "pentest: security penetration scenarios (v2.0.0 gate)",
    "chaos: chaos / failure injection scenarios (v2.0.0 gate)",
    "perf: pytest-benchmark smoke (run explicitly or in scheduled benchmark job)",
]

[tool.mypy]
python_version = "3.14"
strict = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src/syndicateclaw"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
